+++ src/graphics/custom_shape.rs
@@ -13,7 +13,7 @@ pub trait CustomShapePoints {
     /// Gets the total count of points.
     ///
     /// Return the points count
-    fn point_count(&self) -> u32;
+    fn point_count(&self) -> usize;
 
     /// Gets a given point.
     ///
@@ -23,7 +23,7 @@ pub trait CustomShapePoints {
     /// Returns a [`Vector2f`] containing the coordinates.
     ///
     /// [`Vector2f`]: crate::system::Vector2f
-    fn point(&self, point: u32) -> Vector2f;
+    fn point(&self, point: usize) -> Vector2f;
 }
 
 /// A custom textured shape with outline.
@@ -36,13 +36,12 @@ pub struct CustomShape<'s> {
 
 unsafe extern "C" fn get_point_count_callback(obj: *mut c_void) -> usize {
     let shape = obj as *const Box<dyn CustomShapePoints + Send>;
-    let ret = (*shape).point_count();
-    ret as usize
+    (*shape).point_count()
 }
 
 unsafe extern "C" fn get_point_callback(point: usize, obj: *mut c_void) -> sfVector2f {
     let shape = obj as *const Box<dyn CustomShapePoints + Send>;
-    (*shape).point(point.try_into().unwrap())
+    (*shape).point(point)
 }
 
 impl<'s> CustomShape<'s> {