+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 - Methods that used to take `&Rect` now take `Rect` by value
 - `listener::set_up_vector` now Takes `Vector3f` instead of `&Vector3f`
+- Use `usize` instead of `u32` for `CustomShapePoints`
 
 ### Internal improvements
 - Remove needless raw conversion methods for Rect